+++ ScenarioGUI/__init__.py
@@ -1,5 +1,6 @@
-from .global_settings import load as load_config
+import ScenarioGUI.gui_classes.gui_structure_classes as elements
+
 from .gui_classes.gui_combine_window import MainWindow
 from .gui_classes.gui_structure import GuiStructure
 from .translation_csv_to_py import translate_csv_2_class
-import ScenarioGUI.gui_classes.gui_structure_classes as elements
+from .utils import load as load_config