+++ ScenarioGUI/gui_classes/gui_structure_classes/__init__.py
@@ -11,16 +11,18 @@ from .figure_option import FigureOption
 from .filename_box import FileNameBox
 from .flexible_amount_option import FlexibleAmount
 from .float_box import FloatBox
+from .float_box_with_units import FloatBoxWithUnits
 from .function_button import FunctionButton
 from .hint import Hint
 from .int_box import IntBox
+from .int_box_with_units import IntBoxWithUnits
 from .list_box import ListBox
 from .option import Option
 from .page import Page
+from .result_export import ResultExport
 from .result_figure import ResultFigure
 from .result_text import ResultText
 from .text_box import TextBox
-from .result_export import ResultExport
 
 __all__ = [
     "Aim",
@@ -29,9 +31,11 @@ __all__ = [
     "FigureOption",
     "FileNameBox",
     "FloatBox",
+    "FloatBoxWithUnits",
     "FunctionButton",
     "Hint",
     "IntBox",
+    "IntBoxWithUnits",
     "ListBox",
     "Option",
     "Page",