+++ ScenarioGUI/gui_classes/gui_structure_classes/flexible_amount_option.py
@@ -11,8 +11,8 @@ import PySide6.QtWidgets as QtW  # type: ignore
 
 import ScenarioGUI.global_settings as globs
 
+from ...utils import change_font_size, set_default_font
 from .option import Option
-from ...utils import set_default_font, change_font_size
 
 if TYPE_CHECKING:  # pragma: no cover
     from collections.abc import Callable